# BasePlayer modification