--- a/rl_games/common/player.py
+++ b/rl_games/common/player.py
@@ -3,26 +3,32 @@ import gym
 import numpy as np
 import torch
 import copy
+from rl_games.common import vecenv
 from rl_games.common import env_configurations
-from rl_games.algos_torch import  model_builder
+from rl_games.algos_torch import model_builder
+
 
 class BasePlayer(object):
     def __init__(self, params):
         self.config = config = params['config']
         self.load_networks(params)
         self.env_name = self.config['env_name']
+        self.player_config = self.config.get('player', {})
         self.env_config = self.config.get('env_config', {})
+        self.env_config = self.player_config.get('env_config', self.env_config)
         self.env_info = self.config.get('env_info')
         self.clip_actions = config.get('clip_actions', True)
         self.seed = self.env_config.pop('seed', None)
         if self.env_info is None:
-            self.env = self.create_env()
-            self.env_info = env_configurations.get_env_info(self.env)
+            self.env = vecenv.create_vec_env(self.env_name, self.config['num_actors'], **self.env_config)
+            self.env_info = self.env.get_env_info()
+            self.num_agents = 1
         else:
-            self.env = config.get('vec_env')
+            # is there any way 'self.env' can be used in this scenario?
+            # self.env = config.get('vec_env')
+            self.num_agents = self.env_info['agents']
         self.value_size = self.env_info.get('value_size', 1)
         self.action_space = self.env_info['action_space']
-        self.num_agents = self.env_info['agents']
 
         self.observation_space = self.env_info['observation_space']
         if isinstance(self.observation_space, gym.spaces.Dict):