- Player configuration override 

Example:
```
config:
  env_config:
     ... train parameters ... 
  player:
      env_config:
         ... play parameters ... 

```

- Vectorized environment configuration ([Original PR / no activity](https://github.com/Denys88/rl_games/pull/135))

Test plan:

1. Train models for two different scenarios 
python runner.py --train --file rl_games/configs/brax/ppo_ant.yaml 
python runner.py --train --file rl_games/configs/ppo_cartpole.yaml

2. Run model with & without proposed change. (No crash, similar results)
3. Benchmark change (brax/ant)


| Time | Before | After |
| ------------- |  ------------- | ------------- |
| real | 1m32.151s  | 0m42.622s |
| user | 1m21.269s | 0m49.443s |
| sys | 0m31.144s | 0m17.270s |